+++ include/sterm/terminal.hpp
@@ -69,8 +69,9 @@ namespace sterm {
 
       std::string   get_window_title();
       std::string   get_text();
-      bool          copy_font_description(PangoFontDescription **target);
+      gdouble       get_font_scale();
 
+      void          set_font_scale(double font_scale);
       void          set_font_description(PangoFontDescription **font);
       void          insert_text(std::string text);
       void          paste(paste_buffer buffer);