+++ src/sterm/function_handler.cpp
@@ -71,7 +71,7 @@ namespace sterm {
       try {
         func = m_functions.at(command[0]);
       } catch (const std::out_of_range& except) {
-        g_warning("unknown function called: %s", command[0]);
+        g_warning("unknown function called: %s", command[0].c_str());
       }
 
       if ( func != NULL )